backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
androguard
jinja2
aiofiles
pydantic
google-generativeai
groq
